tracking_service/requirements.txt
django==2.1
gunicorn==19.9.0
requests==2.19.1
djangorestframework==3.8.2
django-oauth-toolkit==1.2.0
psycopg2==2.7.5
mongoengine==0.15.3
django-rest-framework-mongoengine==3.3.1

# Test and code quality
pylint==2.1.1
coverage==4.5.1
django-cors-headers==2.4.0